Move files to folder RetrofitModel
